Add Phoenix
This package should re-run the application again without exiting or
closing, for a better user experience and increased complexity, it
should send to the same page as before. But when the refresh button on
the weather_page is triggered, it obviously sends you back to the
home_page, so it was necessary to use the shared_preferences again.